Add transformation for SpeechRate common term
The transformation context in GNU/Linux is as follows:

WPM <-> ESpeak <-> Speech Dispatcher <-> Orca Screen Reader

Orca has a value range of 0..100, where 50 is the default value.
Orca deals with Speech Dispatcher, which has a value range of
-100..100, and having 0 has the default one.
Speech Dispatcher deals with ESpeak, which has (by default) a value
range of 80..390 and 170 is the default value.

There is some ways in which a user can rise the rate up, but there is
not a standard way to do this.