GPII-313: Removing 'fontSize demo' entry from the solutions registry
This solution entry provides the transformation for the fontSize common
term and conflicts with persistentConfiguration's solution entry, which
also provides a transformation for the fontSize common term.

Since the source code for the 'fontSize demo' is not used anymore, it
should also be removed from the native side in a separate pull request
to GPII/android.