GPII-1021: Update some transformations for android and NP sets
The settings handlers are not receiving non-valid (float or boolean)
values from a application specific NP set anymore. As a result of it,
both os_android and talkback1 have been updated according to this.

Some transformations of well-known common terms have been updated to
deal with both float and boolean values that are coming from them.

Also, the "-provisional-general.-provisional-volume" has an updated
value for its "factor" value.