GPII-989: Adding the new WebSockets settings handler
This new settings handler keeps in memory both the available clients
(browsers) and the current available settings for every browser.

The settings handler makes use of web sockets to emit the
onBrowserSettingsChanged signal to the connected clients (if any).
Along with this signal, the settings are being passed to the clients.