GPII-989: Adding a 'browser channel' into the flow manager
Adding a new route to the flowManager, 'browserChannel'.
This is a web socket-based route that allows browsers to connect
with the flowManager via web sockets, in order to make use of the
new WebSockets settings handler.

At every new connection, the browser channel is taking care of:
* Registering the client in the system
* Notifying the client with the current settings
* Making sure that the client will be removed from the system
after being disconnected