GPII-1051: Temporary keyboard a11y
This is a temporary implementation of the keyboard a11y. However this should be refined based on http://oaa-accessibility.org/example/41/, with the exception that the expand/collapse toggles are separated from the items so will have a slightly different interaction. Example single click to expand by clicking on the toggle.