GPII-413: Adding initial mock for dynamic device reporters
At this point it is possible to mock a platform-specific dynamic device
reporter such as the GNU/Linux' packageKit and pass the integration tests.

This first approach mocks the different device reporters on the global path
but it would be better to put them under a well known root path by using
a nameResolver (like the settingsHandlers' mocks).