GPII-2470 - Add script to build and start cloud-based containers
This is a script that can be reused if someone wants to work with Docker containers.

It'll build and start the containers used the cloud-based configuration files.

These containers are then used to run the Production tests.

I have not added this to the Vagrantfile (and made it a separate script) because
the containers are not needed for the normal development workflow. If Docker
containers are adopted as the default workflow in the future, this can be
revisited.

In GPII-2470, I'm only trying to solve the issue where the preferences data
changes and tests need to run against a Preferences Server that has up to
date preferences (instead of https://preferences.gpii.net, which would be
running with a "old" preferences set).

Another approach is to add modify the development configuration files to
change the Kettle port and then start universal again with a different
port. However, port 8081 is hardcoded in too many places in tests and this
might break everything. Besides, it's out of scope for this JIRA to make
all tests dynamicly configurable.